+++ web_v2/resources/limits_with_stats.go
@@ -1,0 +1,22 @@
+package resources
+
+import (
+	"gitlab.com/tokend/horizon/db2/core2"
+	regources "gitlab.com/tokend/regources/generated"
+)
+
+func NewLimitsWithStats(record *core2.LimitsWithStats) regources.LimitsWithStats {
+	return regources.LimitsWithStats{
+		Key: regources.Key{
+			ID:   record.ID,
+			Type: regources.LIMITS_WITH_STATS,
+		},
+		Attributes: regources.LimitsWithStatsAttributes{
+			Statistics: NewStatistics(record.Statistics),
+			Limits:     NewLimits(record.Limits),
+		},
+		Relationships: regources.LimitsWithStatsRelationships{
+			Account: NewAccountKey(record.AccountID).AsRelation(),
+		},
+	}
+}